+++ src/Blitter.Blocks/Behaviors/SpriteBehavior2D.cs
@@ -8,7 +8,7 @@ public abstract class SpriteBehavior2D : Behavior2D
     /// <summary>
     /// Advance the target sprite by one tick.
     /// </summary>
-    public abstract void Update(Sprite2D target, in UpdateContext2D context);
+    public virtual void Update(Sprite2D target, in UpdateContext2D context) {}
 
     /// <summary>
     /// Invoked when the host sprite's <see cref="Sprite2D.HitCircle"/>